Divide story elements to seperate readme into sections
