+++ docs/src/components/Checkbox/Default.tsx
@@ -1,0 +1,78 @@
+import {
+  Checkbox,
+  LoadingIndicator,
+  ThemeProvider,
+  ThemeType,
+  useTheme,
+} from 'dooboo-ui';
+import React, {FC, useState} from 'react';
+
+import {View} from 'react-native';
+import {useFonts} from 'expo-font';
+
+const Component: FC = () => {
+  const [checked, setChecked] = useState<boolean>(false);
+  const {theme} = useTheme();
+
+  const [fontsLoaded] = useFonts({
+    IcoMoon: require('../../assets/doobooui.ttf'),
+  });
+
+  if (!fontsLoaded) return <LoadingIndicator />;
+
+  return (
+    <View
+      style={{
+        backgroundColor: theme.background,
+        flex: 1,
+        alignSelf: 'stretch',
+
+        flexDirection: 'row',
+      }}>
+      <Checkbox checked={checked} onPress={() => setChecked(!checked)} />
+      <View style={{width: 8}} />
+      <Checkbox
+        checked={checked}
+        onPress={() => setChecked(!checked)}
+        type="secondary"
+      />
+      <View style={{width: 8}} />
+      <Checkbox
+        checked={checked}
+        onPress={() => setChecked(!checked)}
+        type="success"
+      />
+      <View style={{width: 8}} />
+      <Checkbox
+        checked={checked}
+        onPress={() => setChecked(!checked)}
+        type="warning"
+      />
+      <View style={{width: 8}} />
+      <Checkbox
+        checked={checked}
+        onPress={() => setChecked(!checked)}
+        type="info"
+      />
+      <View style={{width: 8}} />
+      <Checkbox
+        checked={checked}
+        onPress={() => setChecked(!checked)}
+        type="danger"
+      />
+      <Checkbox
+        checked={checked}
+        onPress={() => setChecked(!checked)}
+        disabled
+      />
+    </View>
+  );
+};
+
+export const Default: FC<{themeType: ThemeType}> = ({themeType}) => {
+  return (
+    <ThemeProvider initialThemeType={themeType}>
+      <Component />
+    </ThemeProvider>
+  );
+};
